KeyframeSelection: Split rescaling parameters for sharpness and flow scores
The rescaled frames used to compute the sharpness and motion scores used
to be the same, with a single parameter to specify the rescale value.
As we may want to use different rescale values depending on whether we are
computing the sharpness or the motion score (or no rescale for one but a
rescale for the other), the existing "rescaledWidth" parameter is split
into two new parameters, "rescaledWidthSharpness" and "rescaledWidthFlow".